pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "volture"
version = "0.1.0"
description = "Kronos RV-IV volatility arbitrage trading agent for Interactive Brokers"
requires-python = ">=3.12"
dependencies = [
    "ib_async>=1.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "torch>=2.1.0",
    "huggingface_hub>=0.25.0",
    "py_vollib>=1.0.1",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[project.scripts]
volture = "volture.cli:main"

[tool.setuptools.packages.find]
include = ["volture*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
